James E Jurach Jr. <muaddib@erf.net> says about his patch:
In XmlRpcServer.java, we find indexOf(".") of methodName to separate the
handlerName from the method name.  This both assumes the handlerName does
not have dots (which may or may not be correct according to the spec...),
but further allows dots to appear in the method name (which is almost
certainly NOT correct, given our mapping of methodName to java method
name).  By chanding that indexOf() to lastIndexOf(), we can better map
handlerName to a package, and methodName to a method name.


I say:

Reading the "Payload format" section of the spec
<http://xmlrpc.org/spec>, I don't see any reason why the handler name
couldn't contain dots.  Josh Lucas concurs.  The spec says:

"The <methodCall> must contain a <methodName> sub-item, a string,
containing the name of the method to be called. The string may only
contain identifier characters, upper and lower-case A-Z, the numeric
characters, 0-9, underscore, dot, colon and slash. It's entirely up to
the server to decide how to interpret the characters in a methodName."

In XmlRpcServer, we choose to interpret the everything after the last
dot as the methodName, and everything before it as the handlerName.


git-svn-id: https://svn.apache.org/repos/asf/webservices/xmlrpc/trunk@233626 13f79535-47bb-0310-9956-ffa450edef68