Moved EMPTY_BYTE_ARRAY back to where Jon had put it in the first place.  Rick Johnston <RRJohnston@symantec.com> indicates that javac (perhaps incorrectly) does not allow it, and provides some information from the JLS:
"In compiling the latest version of XmlRpcServer.java with javac
(1.3.1), I got a compiler error because of the static member
'EMPTY_BYTE_ARRAY' in the inner 'Worker' class.  Section 8.1.2 of the
Java Language Spec 2nd ed. does specifically state that "Inner classes
may not declare static members, unless they are compile-time constant
fields."  Since EMPTY_BYTE_ARRAY is defined as "= new byte[0]", this
is not a compile-time constant and is therefore invalid.
Interestingly, the Jikes compiler does not complain about this
(perhaps because it is also declared as final?).  Alternatively, if
the Worker class was a Nested class but not an Inner class (i.e.,
Worker is 'static' - which doesn't seem appropriate), this would be
legal."


git-svn-id: https://svn.apache.org/repos/asf/webservices/xmlrpc/trunk@233693 13f79535-47bb-0310-9956-ffa450edef68