Work around method signature incompatibilities between version 1.1 and 1.2 of Commons Codec's Base64.encode() and decode() methods.
The 1.1 release had two encode() and decode() methods (each), one that
accepted a byte[] and the other accepted an Object.  In both cases,
Encoder/Decoder Exceptions were thrown if the argument wasn't a
byte[].  In Codec 1.2, both methods still exist, but the one that
accepts a byte[] no longer throws the exception (because the input is
guaranteed to be of type byte[]).


So, I have one option (which kind of sucks), I can cast the byte[] to an
Object in the calls to encode/decode to retain compatibility between codec
1.1 and 1.2.

* src/java/org/apache/xmlrpc/DefaultTypeFactory.java
  createBase64(String): Cast input to the Base64.decode() method from
  byte[] to Object.

* src/java/org/apache/xmlrpc/XmlWriter.java
  writeObject(Object): Cast return value of Base64.encode() from
  Object to byte[].

* src/java/org/apache/xmlrpc/applet/SimpleXmlRpcClient.java
  writeObject(Object, XmlWriter): Cast return value of Base64.encode()
  from Object to byte[].

  Value.characterData(String): Cast input to the Base64.decode()
  method from byte[] to Object.

* src/java/org/apache/xmlrpc/util/HttpUtil.java
  encodeBasicAuthentication(String, String): Cast return value of
  Base64.encode() from Object to byte[], and assure we call the Object
  overload.


Submitted by: Jochen Wiedmann
Reported by:  Ryan Bloom


git-svn-id: https://svn.apache.org/repos/asf/webservices/xmlrpc/trunk@233903 13f79535-47bb-0310-9956-ffa450edef68